+++ windows/vs10/povclr/ColorMap.cpp
@@ -9,7 +9,7 @@
 namespace povclr
 {
 	template<>
-	pov::ColourBlendMapPtr ColorMapDetail::Render<pov::ColourBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	pov::ColourBlendMapPtr ColorMapDetail::Render<pov::ColourBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		// Simple code to eliminate duplicate entries:
 		int p = 0;
@@ -31,13 +31,13 @@ namespace povclr
 
 		pov::GammaCurvePtr blendGammaObj;
 
-		if (double::IsNaN(blendGamma))
+		if (float::IsNaN(blendGamma))
 		{
 			blendGammaObj = pov::PowerLawGammaCurve::GetByDecodingGamma(2.5f);
 		}
 		else
 		{
-			double value = blendGamma;
+			float value = blendGamma;
 			if (value <= 0.0)
 			{
 				value = 0.0;
@@ -55,41 +55,41 @@ namespace povclr
 	}
 
 	template<>
-	pov::GenericPigmentBlendMapPtr ColorMapDetail::Render<pov::GenericPigmentBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	pov::GenericPigmentBlendMapPtr ColorMapDetail::Render<pov::GenericPigmentBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		return ColorMapDetail::Render<pov::ColourBlendMap>(context, tempList, blendMode, blendGamma);
 	}
 
 	template<>
-	pov::PigmentBlendMapPtr ColorMapDetail::Render<pov::PigmentBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	pov::PigmentBlendMapPtr ColorMapDetail::Render<pov::PigmentBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		context->Error("Internal Error: Parse_Colour_Map called for non-colour blend map");
 		return pov::PigmentBlendMapPtr();
 	}
 
 	template<>
-	std::shared_ptr<pov::GenericNormalBlendMap> ColorMapDetail::Render<pov::GenericNormalBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	std::shared_ptr<pov::GenericNormalBlendMap> ColorMapDetail::Render<pov::GenericNormalBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		context->Error("Internal Error: Parse_Colour_Map called for non-colour blend map");
 		return pov::GenericNormalBlendMapPtr();
 	}
 
 	template<>
-	pov::SlopeBlendMapPtr ColorMapDetail::Render<pov::SlopeBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	pov::SlopeBlendMapPtr ColorMapDetail::Render<pov::SlopeBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		context->Error("Internal Error: Parse_Colour_Map called for non-colour blend map");
 		return pov::SlopeBlendMapPtr();
 	}
 
 	template<>
-	pov::NormalBlendMapPtr ColorMapDetail::Render<pov::NormalBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	pov::NormalBlendMapPtr ColorMapDetail::Render<pov::NormalBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		context->Error("Internal Error: Parse_Colour_Map called for non-colour blend map");
 		return pov::NormalBlendMapPtr();
 	}
 
 	template<>
-	pov::TextureBlendMapPtr ColorMapDetail::Render<pov::TextureBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, double blendGamma)
+	pov::TextureBlendMapPtr ColorMapDetail::Render<pov::TextureBlendMap>(gcroot<Context^> context, std::vector<pov::ColourBlendMapEntry>& tempList, int blendMode, float blendGamma)
 	{
 		context->Error("Internal Error: Parse_Colour_Map called for non-colour blend map");
 		return pov::TextureBlendMapPtr();
